Changes to support the new way of setting bluefish properties. now using a,b,c,d for channel selectiong and the keyer property to determine 422, v 4224v interna hardware keyer usage and configuration.
Changes to the casparcg.config file reflect this and provide limited documentation of usage.
aslo changed the executor priority to be highest possible, as this is required for the thread that the DMA happens on.
(building of scatter gather table and actual DMA should never yield to a less important thread)