Changes to support the new way of setting bluefish properties. now using a,b,c,d for channel selectiong and the keyer property to determine 422, v 4224v interna hardware keyer usage and configuration.
Changes to the casparcg.config file reflect this and provide limited documentation of usage.
aslo changed the executor priority to be highest possible, as this is required for the thread that the DMA happens on.
(building of scatter gather table and actual DMA should never yield to a less important thread)


Former-commit-id: 1e06341a1de09a9e037653a090689526cd3a41dc [formerly 1e06341a1de09a9e037653a090689526cd3a41dc [formerly 3f72de1a07476a3eaaae27c98141e86e63a2bd39]]
Former-commit-id: 20b1fe37e933aaa2c9de0a1a24d5d434992fc5cb
Former-commit-id: d24c9b60105cb3113c02da727d91428cd3d84856