Implmented a new system to manage to DMA and presenting of frames when using the bluefish hadrware. This system uses FIFO mode on the hadware which allows for some degree of buffering on the card.
Frames are copied from caspar to a internal Q, a seperate thread then extracts frames from the Q, and does the DMA and present calls.
This system removes our DMA from the core loop of Caspar, and allows up to a full frames time to be used for the DMA.
- Important when using multiple channels.

Moved the Wait for Sync caLl to happen after the Memcopy, this should create a more stable sync time.

No longer need to set the executor thread to high priority, but setting the std::thread that i am using for the DMA to highest priority.

This change results in much better performance when doing multiple streams.

This change required some additional functions to be added to the blue_velvet.h/cpp files too,
to properly support the FIFO playback mode.

Also minor change to tell the Keyer that data is pre-multiplied when we receive it.


Former-commit-id: 47478d1c93b73bd3e145fbf3b34eb0b0ea288462 [formerly 47478d1c93b73bd3e145fbf3b34eb0b0ea288462 [formerly f57e8abe56c9c1afe40c684d0235d1af4eab8e85]]
Former-commit-id: 848049a6a826155c7003f82e4b4a457ebe2b99d7
Former-commit-id: 7060c1168a9fc2218fc07806a47491889f80d689