Upgrade CEF to 3.3239.1705.gf6d6dfc / Chromium 63.0.3239.109.
This brings CasparCG up to a modern version of Chrome, and also changes the structure so that it is easier to upgrade. libcef_dll_wrapper is built locally instead of shipping a binary, as it already has an integrated CMake build system. Upgrading is now just a matter of downloading the latest version (minimal package) and renaming the unpacked directory to dependencies64/cef/windows (or same with /linux); there's no more symlinks or the likes.

Note: this commit does not compile, it just replaces the dependencies. This is to seperate the code changes into a seperate commit to be easily reviewable.

Former-commit-id: 54490a466b3496ec567b783b8528d9fc1352f111 [formerly 5cb5d3e8a39d9ac65f59ddf5ccdcb0a061341059]
Former-commit-id: cccd63def8d30b0e4e30da1bc7397b0bbc623d25