Upgrade CEF to 3.3239.1705.gf6d6dfc / Chromium 63.0.3239.109.
This brings CasparCG up to a modern version of Chrome, and also changes the structure so that it is easier to upgrade. libcef_dll_wrapper is built locally instead of shipping a binary, as it already has an integrated CMake build system. Upgrading is now just a matter of downloading the latest version (minimal package) and renaming the unpacked directory to dependencies64/cef/windows (or same with /linux); there's no more symlinks or the likes.

Note: this commit does not compile, it just replaces the dependencies. This is to seperate the code changes into a seperate commit to be easily reviewable.

Former-commit-id: 24a56e0ffedb3993cbbd9f538ad9c940c35a255d [formerly a7152fdb44ceee7d741abd99d5b76097d0df6893]
Former-commit-id: 16d39875f1b656191ef83591ca3770c1ac17a789