docs: Improve issue reports upon creation for easier handling
More and more issues are reported and sometimes crucial information is missing already from the begining of the issue. By adding this template and asking for information when creating an issue will increaee the quality of issues. It will also minimize the time to give feedback on a certain issue.