This update introduces support for playback and capture of UHD video modes.
The UHD modes are only supported on Kronos Hardware.
This update includes new files in the interop folder and uses the Bluefish provided
BlueVelvetCFuncPtr.h file for loading the function pointers.
This required a change in the files that are included and required for building,
but should be an overall simpler system, and more extensible going forward.
This update also includes a move from using the VIDEO_MODE enum and card property to using
VID_FMT_EXT_ enums and associated card properties which are more powerful, and support the UHD modes.