Prevent single-pipe option from being used with video formats that have an uneven audio cadence.
This is necessary as it is impossible to ensure continued byte synchronisation between the pipe writer and reader under those circumstances without dropping multiple frames every time a single frame is dropped and/or reconnecting the pipe every time a single frame is dropped. So we just prevent it instead.

Video formats with uneven audio candences can be used with separate video/audio pipes with no issue (since the number of bytes written in 1 second is equal, regardless of the cadence sychronisation between write and reader).

Also renamed two variables for consistency.