fix: leave thread_type to any
Just because the codec supports slice-threads, doesn't mean the file
we're opening do. If I understand correctly, for many formats a file
must have been encoded using slices to be able to be decoded in
slices.

By forcing thread-slicing, we make it almost impossible to decode
media that cant be slice-decoded, as it's forcing single-threaded
decoding.