feat: add frame_converter and move consumers across
This removes the `image_data` on `const_frame` and replaces it with an opaque `image_ptr`. Consumers can use the `frame_converter` to convert that into a pixel buffer.
For now this is the same basic texture to buffer copy as before, but will soon be replaced with a format conversion step