fix: remove unnecessary sws scale forcing conversion to YUVA422 (#1668)
There's no need to convert pixel format before we pass frames to
ffmpeg, it should handle conversion as necessary. If your output
pixel format is not YUVA422 ffmpeg would make an additional
conversion. With this change, a single conversion should be
possible.

It may have been required at some point to do parallel slices,
but that should no longer be necessary as ffmpeg/swscale should
be able to do that natively.

Co-authored-by: Jesper Ek <jesper@boffins.se>